change collector delay type

+++ cmd/pitr/collector/collector.go
@@ -30,16 +30,16 @@ type Collector struct {
 }
 
 type Config struct {
-	PXCServiceName string `env:"PXC_SERVICE,required"`
-	PXCUser        string `env:"PXC_USER,required"`
-	PXCPass        string `env:"PXC_PASS,required"`
-	S3Endpoint     string `env:"ENDPOINT" envDefault:"s3.amazonaws.com"`
-	S3AccessKeyID  string `env:"ACCESS_KEY_ID,required"`
-	S3AccessKey    string `env:"SECRET_ACCESS_KEY,required"`
-	S3BucketURL    string `env:"S3_BUCKET_URL,required"`
-	S3Region       string `env:"DEFAULT_REGION,required"`
-	BufferSize     int64  `env:"BUFFER_SIZE"`
-	CollectSpanSec int64  `env:"COLLECT_SPAN_SEC" envDefault:"60"`
+	PXCServiceName string  `env:"PXC_SERVICE,required"`
+	PXCUser        string  `env:"PXC_USER,required"`
+	PXCPass        string  `env:"PXC_PASS,required"`
+	S3Endpoint     string  `env:"ENDPOINT" envDefault:"s3.amazonaws.com"`
+	S3AccessKeyID  string  `env:"ACCESS_KEY_ID,required"`
+	S3AccessKey    string  `env:"SECRET_ACCESS_KEY,required"`
+	S3BucketURL    string  `env:"S3_BUCKET_URL,required"`
+	S3Region       string  `env:"DEFAULT_REGION,required"`
+	BufferSize     int64   `env:"BUFFER_SIZE"`
+	CollectSpanSec float64 `env:"COLLECT_SPAN_SEC" envDefault:"60"`
 }
 
 const (
